% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aethomys_psample}
\alias{aethomys_psample}
\title{Posterior sample of Aethomys trees}
\format{
A \code{multiPhylo} object containing 1000 unrooted trees.
}
\source{
L. Brosseau-Acquaviva et al., in prep.
}
\usage{
aethomys_psample
}
\description{
The posterior sample of cytochrome b trees of a sub-Saharan murid rodent Aethomys.
summarized by the \code{aethomys_tree}.
}
\keyword{datasets}
