% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aethomys_tree}
\alias{aethomys_tree}
\title{Phylogeny of Aethomys}
\format{
A \code{phylo} object containing unrooted tree with 150 ingroup and 3 outgroup tips.
}
\source{
L. Brosseau-Acquaviva et al., in prep.
}
\usage{
aethomys_tree
}
\description{
The maximum clade credibility tree of a sub-Saharan murid rodent Aethomys
based on Bayesian analysis of mitchondrial cytochrome b gene.
}
\keyword{datasets}
