% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcut.R
\name{bcut}
\alias{bcut}
\title{Branchcutting on a single tree.}
\usage{
bcut(
  phy,
  heterosp = NULL,
  consp = NULL,
  rescale = TRUE,
  bw = "mdif",
  kernel = "epanechnikov",
  return.bcut = TRUE
)
}
\arguments{
\item{phy}{an object of class \code{phylo} representing phylogenetic tree with branch lengths.}

\item{heterosp}{a two-column matrix with rows containing tree tips constrained to be heterospecific.}

\item{consp}{a two-column matrix with rows containing tree tips constrained to be conspecific.}

\item{rescale}{logical, whether to rescale branch lengths (default is TRUE).}

\item{bw}{a character string indicating bandwidth function or a number giving the bandwidth;
defaults to \code{mdif}, a bandwith calculation rule designed to the problem in hand.}

\item{kernel}{a character string indicating function used in the kernel density estimation,
defaults to \code{epanechnikov}, but see \code{\link[stats:density]{stats::density()}} for other options.}

\item{return.bcut}{logical, whether to return \code{bcut} object (the full output)
or just a list of delimited species.}
}
\value{
An object of class \code{bcut} or, if \code{return.bcut=FALSE},
a list of character vectors with tree tips assigned to the same species.

The \code{bcut} object is a list with the following components:
\itemize{
\item \code{species} a data frame with columns \code{ID} (tree tip labels)
and \code{Species} (classification to species labelled BCUT1, BCUT2, ...)
\item \code{phy} the input tree (as a \code{phylo} object)
\item \code{score} a numeric vector with importance scores of the tree branches
(listed in the same order as in \code{phy$edge})
\item \code{density} an object of class \code{density} returned by \code{\link[stats:density]{stats::density()}}
in the outlier detection step of the branchcutting algorithm
\item \code{bpoint} the breakpoint, an importance score value separating outliers from the rest
\item \code{backbone} a numeric vector with indices of branches belonging to the interspecific backbone
\item \code{criteria} a data frame with as many rows as the number of tree branches and four components
with logical vectors. The first one (\code{backbone}) indicates which branches belong to the backbone
and the others (\code{outliers}, \code{heterosp} and \code{consp}) upon which criteria they were included.
\item \code{ancestors} a numeric vector with indices of ancestral nodes of the species.
\item \code{nspecies} the number of delimited species
\item \code{rescaled} the tree with rescaled branch length (if the rescaling was applied)
}
}
\description{
\code{bcut} performs branchcutting on a single tree.
}
\examples{
bcspecies <- bcut(dendromus)
}
