% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branchcutting.R
\name{branchcutting}
\alias{branchcutting}
\title{Branchcutting species delimitation.}
\usage{
branchcutting(
  phy,
  psample,
  heterosp = NULL,
  consp = NULL,
  rescale = TRUE,
  bw = "mdif",
  kernel = "epanechnikov",
  outgroup = NULL
)
}
\arguments{
\item{phy}{Phylogenetic tree with branch lengths supplied as an object of class \code{phylo},
a character string in newick format or a name of file with the tree in newick or nexus format.}

\item{psample}{A posterior sample of trees as a list of \code{phylo} obbjects, \code{multiPhylo} object,
or a name of file with the trees in newick or nexus format.}

\item{heterosp}{A two-column matrix with rows containing tree tips constrained to be heterospecific
or a name of tab-delimited file without header containing such matrix.}

\item{consp}{A two-column matrix with rows containing tree tips constrained to be conspecific.
or a name of tab-delimited file without header containing such matrix.}

\item{rescale}{Logical, whether to rescale branch lengths (default is TRUE).}

\item{bw}{A character string indicating bandwidth function or a number giving the bandwidth;
defaults to \code{mdif}, a bandwith calculation rule designed to the problem in hand.}

\item{kernel}{A character string indicating function used in the kernel density estimation,
defaults to \code{epanechnikov}, but see \code{\link[stats:density]{stats::density()}} for other options.}

\item{outgroup}{A character vector with outgroup names or their unique identifier(s).}
}
\value{
An object of class \code{bcut} or, if only \code{psample} is supplied,
a list with components \code{consdelim} (data frame with the consensus species delimitation) and
\code{psample} (a list containing information about delimited species and their relative supports).
}
\description{
The function \code{branchcutting} delimits species on a single tree and/or
a posterior sample of trees using branch-cutting algorithm.
}
\examples{
hetspmat <- matrix(c("TZ30480", "TZ30767"), 1, 2)
conspmat <- matrix(c("RS1629", "TZ30767"), 1, 2)
bcspecies <- branchcutting(phy=aethomys_tree, psample=aethomys_psample, heterosp=hetspmat, consp=conspmat, outgroup="outgroup")
}
