% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcut.R
\name{calculate_importance}
\alias{calculate_importance}
\title{Importance scores.}
\usage{
calculate_importance(phy)
}
\arguments{
\item{phy}{An object of class \code{phylo}.}
}
\value{
A numeric vector with importance scores of the tree branches.
}
\description{
Calculates importance score for every branch in the tree.
}
