% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dendromus}
\alias{dendromus}
\title{Phylogeny of Dendromus}
\format{
A \code{phylo} object containing rooted tree with 45 tips and no outgroup.
}
\source{
G. Mulualem et al., in prep.
}
\usage{
dendromus
}
\description{
A phylogenetic tree of Ethiopian African climbing mice (Dendromus)
based on ML analysis of concatenated ddRAD SNPs.
}
\keyword{datasets}
