% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export.bcut}
\alias{export.bcut}
\title{Export of branchcutting delimitation.}
\usage{
\method{export}{bcut}(
  x,
  file = NULL,
  return = TRUE,
  taxonomy = NULL,
  col.names = c("ID", "Species")
)
}
\arguments{
\item{file}{name for the tab-delimited output file.}

\item{return}{logical, whether to return result as a data frame.}

\item{taxonomy}{a data frame with an existing classification, informing about species names.}

\item{col.names}{preferred column names (1st must refer to tree tips, 2nd to the species).}

\item{bc}{an object of class \code{bcut}.}
}
\description{
\code{export} method for class \code{bcut}.
}
