% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{Exporting species delimitations.}
\usage{
export(x, ...)
}
\description{
Generic function, intended for exporting species delimitations into plain text tables and/or data frames.
}
