% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcut.R
\name{find_crowns}
\alias{find_crowns}
\title{Tree crowns.}
\usage{
find_crowns(phy, tips)
}
\arguments{
\item{phy}{An object of class \code{phylo}.}

\item{tips}{A two-column matrix with pairs of tips (indicated by names or numbers).}
}
\value{
A logical vector indicating whether the tree branches are in any crown.
}
\description{
Finds tree crowns, meaning a set of subtrees consisting of the most recent
common ancestors of the specified tips and all their offspring.
}
