% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcut.R
\name{find_outliers}
\alias{find_outliers}
\title{Detection of outlier branches.}
\usage{
find_outliers(score, bw, kernel, backbone)
}
\arguments{
\item{score}{a numeric vector with importance scores of the tree branches.}

\item{bw}{the bandwidth parameter of the kernel density function.}

\item{kernel}{a character string indicating the kernel density function.}

\item{backbone}{logical vector indicating branches constrained to be in the interspecific backbone.}
}
\value{
A list with the components \code{outliers} (logical vector indicating outlying branches),
\code{density} (an object with class \code{density} returned by the \code{\link[stats:density]{stats::density()}}) and
\code{bpoint} (the breakpoint, an importance score value separating outliers from the rest).
}
\description{
Finds branches with outlying (very high) importance scores.
}
