% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcut.R
\name{find_periphery}
\alias{find_periphery}
\title{Periphery of the tree.}
\usage{
find_periphery(zero, first, phy)
}
\arguments{
\item{zero}{an integer giving the number of the ancestral node (starting point on the path to periphery).}

\item{first}{an integer vector with numbers of the first other nodes on the paths to the periphery.
It specifies the direction in which to search for the specific segments of the the periphery.}

\item{phy}{An object of class \code{phylo}.}
}
\value{
A vector with numbers of tips belonging to the specified segment(s) of periphery.
}
\description{
Finds peripheral tips in given direction on an unrooted tree.
}
