% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsupport.R
\name{getsupport}
\alias{getsupport}
\title{Branchcutting support of species.}
\usage{
getsupport(species, bc)
}
\arguments{
\item{species}{Species for which support is to be assessed. Either a two-column matrix or data frame
(1st column tip labels, 2nd column species) or a (list of) vector(s) with tip labels
defining the species.}

\item{bc}{An object of class \code{bcut}.}
}
\value{
A numeric vector with supports for species defined in \code{species}.
}
\description{
Extracts support for specified species from branchcutting results.
}
\details{
The input \code{bcut} object must include component \code{psample}.
}
