% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative.R
\name{iterative}
\alias{iterative}
\title{Iterative branchcutting.}
\usage{
iterative(
  phy,
  heterosp = NULL,
  consp = NULL,
  rescale = TRUE,
  bw = "mdif",
  kernel = "epanechnikov",
  outgroup = NULL
)
}
\arguments{
\item{phy}{phylogenetic tree with branch lengths supplied as an object of class \code{phylo},
a character string in newick format or a name of file with the tree in newick or nexus format.}

\item{heterosp}{as in \link{branchcutting}, initial heterospecific constraint.}

\item{consp}{as in \link{branchcutting}, initial conspecific constraint.}

\item{rescale}{logical, whether to rescale branch lengths (default is TRUE).}

\item{bw}{a character string indicating bandwidth function or a number giving the bandwidth;
defaults to \code{mdif}, a bandwith calculation rule designed to the problem in hand.}

\item{kernel}{a character string indicating function used in the kernel density estimation,
defaults to \code{epanechnikov}, but see \code{\link[stats:density]{stats::density()}} for other options.}

\item{outgroup}{a character vector with outgroup names or their unique identifier(s).}
}
\value{
a list with two components: \code{iterations} containing a list of \code{bcut} objects
(results from subsequent iterations) and \code{species}, a data frame with the final delimitation.
}
\description{
Iterative application of the branchcutting species delimitation to the same tree
using delimitation from a previous run as heterospecific constraint to the next one.
}
