% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative.R
\name{prepare_constraints}
\alias{prepare_constraints}
\title{Preparing constraints.}
\usage{
prepare_constraints(phy, spdelim, type = "heterospecific")
}
\arguments{
\item{phy}{phylogenetic tree with branch lengths supplied as an object of class \code{phylo},
a character string in newick format or a name of file with the tree in newick or nexus format.}

\item{spdelim}{an object of class \code{bcut} or a data frame or matrix
with tree tips in the first column and their classification into species in the second
or a name of tab-delimited file containing such data.}

\item{type}{character string specifies the type of constraint, either "heterospecific" (default)
or "conspecific" or their unambiguous abbreviation.}
}
\value{
a two-column matrix specifying the constraints.
}
\description{
Prepares constraints for species delimitation using branchcutting algorithm.
}
