% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ptp.R
\name{read_ptp}
\alias{read_ptp}
\title{Reads files produced by PTP / mPTP.}
\usage{
read_ptp(file)
}
\arguments{
\item{file}{character, name of the file.}
}
\value{
An object of class \code{spdelim}.
}
\description{
The function reads files with results of PTP / mPTP species delimitation.
}
