% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcut.R
\name{rescale_branchlengths}
\alias{rescale_branchlengths}
\title{Rescaling of branch lengths.}
\usage{
rescale_branchlengths(phy)
}
\arguments{
\item{phy}{An object of class \code{phylo}.}
}
\value{
An object of class \code{phylo} with rescaled branch lengths.
}
\description{
Rescales branch lengths according to a proper Laplacian graph transform.
}
