% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootanddrop.R
\name{rootanddrop}
\alias{rootanddrop}
\title{Roots the tree and drops the outgroup.}
\usage{
rootanddrop(phy, outgroup = "outgroup")
}
\arguments{
\item{phy}{An object of class \code{phylo} or \code{multiPhylo}.}

\item{outgroup}{A character vector with outgroup names or their unique identifier(s).}
}
\value{
A rooted tree without outgroup(s).
}
\description{
The function performs outgroup-rooting of the tree and then drops the outgroup(s) out.
}
\examples{
rootanddrop(phy=aethomys_tree, outgroup="outgroup")
}
