% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcut.R
\name{find_backbone}
\alias{find_backbone}
\title{Tree backbone.}
\usage{
find_backbone(phy, nodes)
}
\arguments{
\item{phy}{an object of class \code{phylo}.}

\item{nodes}{a numeric vector with node numbers.}
}
\value{
A logical vector indicating whether the tree branches are in the backbone.
}
\description{
Finds basal / central backbone of a tree anchored by specified nodes.
}
