% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsupport.R
\name{getconsensus}
\alias{getconsensus}
\title{Consensus species delimitation.}
\usage{
getconsensus(bc)
}
\arguments{
\item{bc}{An object of class \code{bcut}.}
}
\description{
Extracts onsensus species delimitation from a branchcutting analysis.
}
\details{
The input \code{bcut} object must include component \code{psample}.
}
